requirements.txt
fastapi[standard]
pymongo 
python-dotenv
python-multipart
"uvicorn[standard]" 
pydantic 
httpx 
motor